MSI: Migrate Single Stock Data From CatalogInventory to MSI SourceItems
-- slight refactoring

+++ app/code/Magento/InventoryApi/Test/_files/products_rollback.php
@@ -5,11 +5,13 @@
  */
 declare(strict_types=1);
 
+use Magento\Framework\Api\SearchCriteriaBuilder;
 use Magento\Catalog\Api\ProductRepositoryInterface;
 use Magento\CatalogInventory\Api\StockStatusCriteriaInterfaceFactory;
 use Magento\CatalogInventory\Api\StockStatusRepositoryInterface;
 use Magento\Framework\Registry;
 use Magento\TestFramework\Helper\Bootstrap;
+use Magento\Catalog\Api\Data\ProductInterface;
 
 $objectManager = Bootstrap::getObjectManager();
 /** @var ProductRepositoryInterface $productRepository */
@@ -25,18 +27,31 @@ $currentArea = $registry->registry('isSecureArea');
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', true);
 
-for ($i = 1; $i <= 3; $i++) {
-    $product = $productRepository->get('SKU-' . $i);
-    /** @var \Magento\CatalogInventory\Api\StockStatusCriteriaInterfaceFactory $stockStatusCriteriaFactory **/
-    $criteria = $stockStatusCriteriaFactory->create();
-    $criteria->setProductsFilter($product->getId());
+/** @var SearchCriteriaBuilder $searchCriteriaBuilder */
+$searchCriteriaBuilder = Bootstrap::getObjectManager()->get(SearchCriteriaBuilder::class);
+$searchCriteria = $searchCriteriaBuilder->addFilter(
+    ProductInterface::SKU,
+    ['SKU-1', 'SKU-2', 'SKU-3'],
+    'in'
+)->create();
+$products = $productRepository->getList($searchCriteria)->getItems();
 
-    $result = $stockStatusRepository->getList($criteria);
-    $stockStatus = current($result->getItems());
-    $stockStatusRepository->delete($stockStatus);
+/**
+ * Tests which are wrapped with MySQL transaction clear all data by transaction rollback.
+ * In that case there is "if" which checks that SKU1, SKU2 and SKU3 still exists in database.
+ */
+if (!empty($products)) {
+    foreach ($products as $product) {
+        /** @var \Magento\CatalogInventory\Api\StockStatusCriteriaInterfaceFactory $stockStatusCriteriaFactory */
+        $criteria = $stockStatusCriteriaFactory->create();
+        $criteria->setProductsFilter($product->getId());
 
-    $productRepository->deleteById('SKU-' . $i);
+        $result = $stockStatusRepository->getList($criteria);
+        $stockStatus = current($result->getItems());
+        $stockStatusRepository->delete($stockStatus);
+
+        $productRepository->delete($product);
+    }
 }
-
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', $currentArea);
